requirements.txt
# Discord
discord.py
boto